Subject: flourish & .dropcap styles are decoupled, remove avif due to safari not showing them. Reason for Commit: Floursh was attached to container not first nested p, and so pages starting with an h2 for example were a bit strange. Also avif were not showing up in Safari.
Summary:
- Decoupled ::before psudo elements between .dropcap and .flourish
- Improved method of loading properties via define:vars
- Removed avif due to strange safari behavior.

Notes:
I cannot find a registered bug but in safari 17.5 the avif was being chosen from within the image-set, but not rendering. Removing the type() and adding a 1x resolution flag did not effect things. Simply assigning the avif as a background-image: url() also didn't work. Caniuse shows these as supported, but this isn't working, while webp, png etc work as expected.
https://caniuse.com/?search=image-set
https://caniuse.com/?search=avif