fix(app): Fix bug where default speed wasn't selected on load
This was a combo of checking both radio.checked and radio.hasAttribute('checked') as for some reason the results were not consistant for radio.checked. We also had to do the same with css, as inputs with 'checked' attributes were not consistantly registering, infact only the last one on the page was. I belive this is a race condition, or perhaps the componetns are not all mounted to register a checked state to js. We will need to observe if this solution is sufficiant for our use case.