fix(app): Fix playback speed with mutiple players
We have changed the UUID to a suffix rather then prefix for read-ability. Also I believe we have narrowed down the playback speed issue to a race condition with the browser's reading of the DOM and our re-naming idfor and name attributes in the constructor. The only ways I can see arround this would be either not use a slot for our markup generation and do it all on the clent in template literals (not great), or to move uuid generation over to the Astro component on build.The later approach is probably preferable, however we need to decide what if any control we give to the user.