feat(app): Auto routing, and improved accessibility
Added ID attr which us used on id and aria-label, also a currentPage attr which can be manualy used to generate paths from the Astro.url object, or as a fallback a crumbs array to stucture your own breadcrumbs. Text is automatically generated from hypen seperated string in the path, using '/' as the url delimiter.