fix(app): Correct links to sites
Interesting behavior with links. You can use <a href="./bla"> to link to other files in the same directory. HOWEVER if the url is https://your.domain/foo (no trialing slash) it wont work. It HAS TO have a trailing slash. eg https://your.domain/foo/. To get arround include the root in the link href. <a href="../foo/bla">