feat(app): Refine strategy for canonical links
Refine strategy for canonical links, and identifying current domain. We now use link rel="alternate" on the the cannonical domain to point to the mirror, and on the mirror, use link rel="canonical" to point to the canonical domain. In our case we are using .github.io as the canonical domain, as it will live the longest.